requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
django-filter==24.2
simplejwt
coverage
drf-yasg
django-cors-headers
requests